Temperature sensor ugly conversion, see #21

+++ src/core/analog_read.h
@@ -12,12 +12,13 @@
 #define MAX_TEMP0 50
 #define MAX_TEMP1 50
 
+
 extern volatile int16_t p;
 extern volatile int16_t p_peak;
 extern volatile int16_t cycle_p_peak;
 
-extern volatile uint8_t temp0;
-extern volatile uint8_t temp1;
+extern volatile int16_t temp0;
+extern volatile int16_t temp1;
 
 void init_analog_read();
